When employees are linked to task cannot delete

+++ lib/controllers/EmployeesController.dart
@@ -1,6 +1,7 @@
 import 'package:admin/models/EmployeeModel.dart';
 import 'package:admin/services/google_sheets_api.dart';
 import 'package:mvc_pattern/mvc_pattern.dart';
+import 'package:admin/models/TaskModel.dart';
 
 class EmployeesController extends ControllerMVC 
 {
@@ -18,7 +19,14 @@ class EmployeesController extends ControllerMVC
 
   void remove(Employee employee)
   {
-    EmployeeModel.remove(employee);
-    GoogleSheetsApi.deleteEmployee(employee.id!);
+    if (!TaskModel.isAssigned(employee))
+    {
+      EmployeeModel.remove(employee);
+      GoogleSheetsApi.deleteEmployee(employee.id!);
+    }
+    else
+    {
+      print('Cannot delete employee because its assigned to a task...');
+    }
   }
 }
